+++ expression/column.go
@@ -147,9 +147,7 @@ type Column struct {
 	TblName     model.CIStr
 	RetType     *types.FieldType
 	ID          int64
-	// Position means the position of this column that appears in the select fields.
-	// e.g. SELECT name as id , 1 - id as id , 1 + name as id, name as id from src having id = 1;
-	// There are four ids in the same schema, so you can't identify the column through the FromID and ColName.
+	// Position is the unique id of this column.
 	Position int
 	// IsAggOrSubq means if this column is referenced to a Aggregation column or a Subquery column.
 	// If so, this column's name will be the plain sql text.
